requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-dotenv==1.2.1
google-genai==1.56.0
pinecone==4.0.0
pypdf==6.5.0
